if not 3D then do not extract

+++ bash_scripts/aio.sh
@@ -10,8 +10,13 @@ do
   
   if [[ $iso != "" ]]; then
     echo $iso
-    dotnet /home/ftpuser/bdextract/BDExtractor.dll -p $iso -o $blurayfolderitem
-    rm $iso
+
+    if [[ $iso != *"3D"* ]];then
+      dotnet /home/ftpuser/bdextract/BDExtractor.dll -p $iso -o $blurayfolderitem
+      rm $iso
+    else
+      blurayfolderitem=$iso
+    fi
   fi  
 
   bash /home/bdscript_with_ss.sh $blurayfolderitem
